requirements.txt
piexif
fire
tqdm
pyyaml
